+++ sql/core/src/test/scala/org/apache/spark/sql/execution/command/v2/DropTableSuite.scala
@@ -49,26 +49,4 @@ class DropTableSuite extends command.DropTableSuiteBase with CommandSuiteBase {
         Seq.empty)
     }
   }
-
-  test("SPARK-33305: DROP TABLE should also invalidate cache") {
-    val t = s"$catalog.ns.tbl"
-    val view = "view"
-    withNamespace(s"$catalog.ns") {
-      sql(s"CREATE NAMESPACE $catalog.ns")
-      withTempView(view, "source") {
-        val df = spark.createDataFrame(Seq((1L, "a"), (2L, "b"), (3L, "c"))).toDF("id", "data")
-        df.createOrReplaceTempView("source")
-        sql(s"CREATE TABLE $t $defaultUsing AS SELECT id, data FROM source")
-        sql(s"CACHE TABLE $view AS SELECT id FROM $t")
-        checkAnswer(sql(s"SELECT * FROM $t"), spark.table("source").collect())
-        checkAnswer(
-          sql(s"SELECT * FROM $view"),
-          spark.table("source").select("id").collect())
-
-        assert(!spark.sharedState.cacheManager.lookupCachedData(spark.table(view)).isEmpty)
-        sql(s"DROP TABLE $t")
-        assert(spark.sharedState.cacheManager.lookupCachedData(spark.table(view)).isEmpty)
-      }
-    }
-  }
 }
